FEATURE: Changed config system to use an xmmsv dictionary structure, holding xmmsv nodes
Old API was preserved except for the change in xmms_config_property_lookup_get_string and
xmms_config_property_get_string, which now return a gchar* instead of const gchar *.
The caller is responsible for freeing this. Modified all plugins that called this function
to adhere to the new behavior.
Bumped XMMS_CONFIG_VERSION to 3, because a new attribute "type" for properties was added to the xml representation